test(Area): update test cases
